requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
python-dotenv
genai